Superluous DAO removed
